fix some consistency pitfalls
no duplication-check mechanism is atomic. In addition to checking before
adding, we shall use an unique `_id` to avoid the rare condition that
two requests simultaneously pass the duplication check.

This only applies to those `create` functions where we require some
field to not duplicate inside the database. If some function doesn't
need to check for duplication, then there is no such pitfall.